+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-rc2] -- unreleased
+
+## Changed
+
+- clean up, code organization
+
+## Added
+
+- update via background job
+- console command to update manually
+
 ## [1.0.0-rc1] -- unreleased
 
 ## Added